Helper\CriteraHelper.php moved to Criteria\ContentType.php
